backend/requirements.txt
pandas==2.2.3
Flask==3.0.3
Flask-Cors==5.0.0
python-dotenv==1.0.1
requests==2.32.3
spotipy==2.24.0